backend: refactor constants and disable caching
Not sure if this will work, but apparently not calling noCache will
cause "undefined" to be rolled into the generated files during building.
We want runtime environment variables, so we should not do this.

Also did some refactoring because importing all of the needed constants
in each file was getting a bit too long and tedious.